rename: util 폴더 구조 수정, 관련 import 위치 변경
